extras: Factored out module to run calculations in the background
