+++ klippy/extras/background_process.py
@@ -1,0 +1,43 @@
+# Background execution of calculations in a separate process
+#
+# Copyright (C) 2018-2022  Kevin O'Connor <kevin@koconnor.net>
+# Copyright (C) 2022  Dmitry Butyugin <dmbutyugin@google.com>
+#
+# This file may be distributed under the terms of the GNU GPLv3 license.
+import multiprocessing, traceback
+import queuelogger
+
+def background_process_exec(printer, method, args):
+    if printer is None:
+        return method(*args)
+    parent_conn, child_conn = multiprocessing.Pipe()
+    def wrapper():
+        queuelogger.clear_bg_logging()
+        try:
+            res = method(*args)
+        except:
+            child_conn.send((True, traceback.format_exc()))
+            child_conn.close()
+            return
+        child_conn.send((False, res))
+        child_conn.close()
+    # Start a process to perform the calculation
+    calc_proc = multiprocessing.Process(target=wrapper)
+    calc_proc.daemon = True
+    calc_proc.start()
+    # Wait for the process to finish
+    reactor = printer.get_reactor()
+    gcode = printer.lookup_object("gcode")
+    eventtime = last_report_time = reactor.monotonic()
+    while calc_proc.is_alive():
+        if eventtime > last_report_time + 5.:
+            last_report_time = eventtime
+            gcode.respond_info("Wait for calculations..", log=False)
+        eventtime = reactor.pause(eventtime + .1)
+    # Return results
+    is_err, res = parent_conn.recv()
+    if is_err:
+        raise printer.command_error("Error in remote calculation: %s" % (res,))
+    calc_proc.join()
+    parent_conn.close()
+    return res
