+++ klippy/extras/shaper_calibrate.py
@@ -3,7 +3,8 @@
 # Copyright (C) 2020  Dmitry Butyugin <dmbutyugin@google.com>
 #
 # This file may be distributed under the terms of the GNU GPLv3 license.
-import collections, importlib, logging, math, multiprocessing, traceback
+import collections, functools, importlib, logging, math, multiprocessing
+background_process = importlib.import_module('.background_process', 'extras')
 shaper_defs = importlib.import_module('.shaper_defs', 'extras')
 
 MIN_FREQ = 5.
@@ -61,6 +62,8 @@ class ShaperCalibrate:
     def __init__(self, printer):
         self.printer = printer
         self.error = printer.command_error if printer else Exception
+        self.background_process_exec = functools.partial(
+                background_process.background_process_exec, printer)
         try:
             self.numpy = importlib.import_module('numpy')
         except ImportError:
@@ -68,42 +71,6 @@ class ShaperCalibrate:
                     "Failed to import `numpy` module, make sure it was "
                     "installed via `~/klippy-env/bin/pip install` (refer to "
                     "docs/Measuring_Resonances.md for more details).")
-
-    def background_process_exec(self, method, args):
-        if self.printer is None:
-            return method(*args)
-        import queuelogger
-        parent_conn, child_conn = multiprocessing.Pipe()
-        def wrapper():
-            queuelogger.clear_bg_logging()
-            try:
-                res = method(*args)
-            except:
-                child_conn.send((True, traceback.format_exc()))
-                child_conn.close()
-                return
-            child_conn.send((False, res))
-            child_conn.close()
-        # Start a process to perform the calculation
-        calc_proc = multiprocessing.Process(target=wrapper)
-        calc_proc.daemon = True
-        calc_proc.start()
-        # Wait for the process to finish
-        reactor = self.printer.get_reactor()
-        gcode = self.printer.lookup_object("gcode")
-        eventtime = last_report_time = reactor.monotonic()
-        while calc_proc.is_alive():
-            if eventtime > last_report_time + 5.:
-                last_report_time = eventtime
-                gcode.respond_info("Wait for calculations..", log=False)
-            eventtime = reactor.pause(eventtime + .1)
-        # Return results
-        is_err, res = parent_conn.recv()
-        if is_err:
-            raise self.error("Error in remote calculation: %s" % (res,))
-        calc_proc.join()
-        parent_conn.close()
-        return res
 
     def _split_into_windows(self, x, window_size, overlap):
         # Memory-efficient algorithm to split an input 'x' into a series
